removed user api, and moved resume upload route
