+++ components/applicationSteps/ApplicationStep.tsx
@@ -1,15 +1,15 @@
 import * as React from "react";
 
 type Props = {
-  user: any;
+  profile: any;
 };
 
 const ApplicationStep: React.FunctionComponent<Props> = props => {
-  const { user } = props;
+  const { profile } = props;
 
   return (
     <div>
-      <h1>Hello {user.displayName}</h1>
+      <h1>Hello {profile.displayName}</h1>
 
       <p>
         This is a component that will represent the user's application. There
